Changed the print format from long to int
git-svn-id: http://internalbrr.cr.usgs.gov/svn_MOWS/PRMS_SVN/mmf/trunk@5228 6316d359-22eb-4c0b-9c0f-3bdc8ec1e293

+++ print_params.c
@@ -171,9 +171,12 @@ int print_params (void) {
 
     switch(param->type) {
        case M_LONG:
-          (void)fprintf (param_file, "Max       : %ld\n", *(long *)(param->max));
-          (void)fprintf (param_file, "Min       : %ld\n", *(long *)(param->min));
-          (void)fprintf (param_file, "Default   : %ld\n", *(long *)(param->def));
+//          (void)fprintf (param_file, "Max       : %ld\n", *(long *)(param->max));
+//          (void)fprintf (param_file, "Min       : %ld\n", *(long *)(param->min));
+//          (void)fprintf (param_file, "Default   : %ld\n", *(long *)(param->def));
+          (void)fprintf (param_file, "Max       : %d\n", *(int *)(param->max));
+          (void)fprintf (param_file, "Min       : %d\n", *(int *)(param->min));
+          (void)fprintf (param_file, "Default   : %d\n", *(int *)(param->def));
           break;
 
        case M_FLOAT:
