+++ save_params.c
@@ -194,7 +194,8 @@ static void write_parameters (FILE *param_file, int writeAllParams) {
 	long i,j;
 	double	*dvalptr;
 	float	*fvalptr;
-	long	*lvalptr;
+//	long	*lvalptr;
+	int	*lvalptr;
 /*
 * Write out parameter values and description if any.
 */
@@ -267,13 +268,16 @@ static void write_parameters (FILE *param_file, int writeAllParams) {
 
 				case M_LONG:
 					if (writeAllParams) {
-						lvalptr = (long *) param->value;
+//						lvalptr = (long *) param->value;
+						lvalptr = (int *) param->value;
 					} else {
-						lvalptr = (long *) (param->references[0]);
+//						lvalptr = (long *) (param->references[0]);
+						lvalptr = (int *) (param->references[0]);
 					}
 
 					for (j = 0; j < param->size; j++) {
-						(void)fprintf(param_file, "%ld\n", *lvalptr);
+//						(void)fprintf(param_file, "%ld\n", *lvalptr);
+						(void)fprintf(param_file, "%d\n", *lvalptr);
 						lvalptr++;
 						if (param->value_desc[j]) {
 						  while ((ptr = strchr (param->value_desc[j], '\n'))) {
